Refactor Solr bioentities core code to solr.bioentities package
