+++ base/src/main/java/uk/ac/ebi/atlas/experimentimport/analyticsindex/BioentityPropertiesDao.java
@@ -2,7 +2,7 @@ package uk.ac.ebi.atlas.experimentimport.analyticsindex;
 
 import uk.ac.ebi.atlas.model.analyticsindex.ExperimentDataPoint;
 import uk.ac.ebi.atlas.model.experiment.baseline.BioentityPropertyName;
-import uk.ac.ebi.atlas.solr.query.BioentitiesSolrClient;
+import uk.ac.ebi.atlas.solr.bioentities.query.BioentitiesSolrClient;
 import com.google.common.collect.ImmutableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
