+++ base/src/main/java/uk/ac/ebi/atlas/model/resource/BioentityPropertyFile.java
@@ -2,7 +2,7 @@ package uk.ac.ebi.atlas.model.resource;
 
 import com.google.common.base.Preconditions;
 import org.apache.commons.lang.StringUtils;
-import uk.ac.ebi.atlas.solr.BioentityProperty;
+import uk.ac.ebi.atlas.solr.bioentities.BioentityProperty;
 import uk.ac.ebi.atlas.species.Species;
 
 import java.nio.file.Path;
