+++ base/src/main/java/uk/ac/ebi/atlas/solr/bioentities/admin/BioentityPropertiesSource.java
@@ -1,10 +1,10 @@
-package uk.ac.ebi.atlas.solr.admin;
+package uk.ac.ebi.atlas.solr.bioentities.admin;
 
 import org.apache.commons.lang.ArrayUtils;
 import org.springframework.beans.factory.annotation.Value;
 import uk.ac.ebi.atlas.model.experiment.baseline.BioentityPropertyName;
 import uk.ac.ebi.atlas.model.resource.BioentityPropertyFile;
-import uk.ac.ebi.atlas.solr.BioentityProperty;
+import uk.ac.ebi.atlas.solr.bioentities.BioentityProperty;
 import uk.ac.ebi.atlas.species.Species;
 import uk.ac.ebi.atlas.species.SpeciesFactory;
