+++ base/src/main/java/uk/ac/ebi/atlas/solr/bioentities/query/GeneQueryResponse.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.atlas.solr.query;
+package uk.ac.ebi.atlas.solr.bioentities.query;
 
 import com.google.common.collect.LinkedHashMultimap;
 import com.google.common.collect.Multimap;
