+++ base/src/main/java/uk/ac/ebi/atlas/solr/bioentities/query/SolrQueryService.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.atlas.solr.query;
+package uk.ac.ebi.atlas.solr.bioentities.query;
 
 import uk.ac.ebi.atlas.search.SemanticQuery;
 import uk.ac.ebi.atlas.search.SemanticQueryTerm;
@@ -6,7 +6,7 @@ import com.google.common.base.Stopwatch;
 import org.apache.solr.client.solrj.SolrQuery;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.ac.ebi.atlas.solr.query.builders.BioentityIdentifierQueryBuilder;
+import uk.ac.ebi.atlas.solr.bioentities.query.builders.BioentityIdentifierQueryBuilder;
 import uk.ac.ebi.atlas.species.Species;
 import uk.ac.ebi.atlas.web.GenesNotFoundException;
