+++ base/src/test/java/uk/ac/ebi/atlas/bioentity/properties/BioentityPropertyDaoTest.java
@@ -10,7 +10,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.ac.ebi.atlas.controllers.BioentityNotFoundException;
 import uk.ac.ebi.atlas.model.experiment.baseline.BioentityPropertyName;
-import uk.ac.ebi.atlas.solr.query.BioentitiesSolrClient;
+import uk.ac.ebi.atlas.solr.bioentities.query.BioentitiesSolrClient;
 
 import java.util.HashMap;
 import java.util.Set;
