+++ base/src/test/java/uk/ac/ebi/atlas/solr/bioentities/query/GeneQueryResponseTest.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.atlas.solr.query;
+package uk.ac.ebi.atlas.solr.bioentities.query;
 
 import com.google.common.collect.Sets;
 import org.junit.Before;
