+++ base/src/test/resources/configuration.properties
@@ -3,7 +3,7 @@ experimentsFilesLocation = ${data.files.location}/gxa/
 # uk.ac.ebi.atlas.species.SpeciesPropertiesDao
 speciesPropertiesLocation = ${data.files.location}/species/
 
-# uk.ac.ebi.atlas.solr.admin.BioentityIndexAdmin
+# uk.ac.ebi.atlas.solr.bioentities.admin.BioentityIndexAdmin
 bioentity.properties = ${data.files.location}/bioentity_properties/
 
 # Ontology traders
