+++ gxa/src/test/java/uk/ac/ebi/atlas/controllers/rest/AutoCompleteControllerTest.java
@@ -9,7 +9,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import uk.ac.ebi.atlas.solr.query.SolrBioentitiesSuggesterService;
+import uk.ac.ebi.atlas.solr.bioentities.query.SolrBioentitiesSuggesterService;
 import uk.ac.ebi.atlas.search.SemanticQueryTerm;
 import uk.ac.ebi.atlas.species.Species;
 import uk.ac.ebi.atlas.species.SpeciesFactory;
