+++ sc/src/main/resources/configuration.properties
@@ -5,7 +5,7 @@ experimentsFilesLocation = ${data.files.location}/scxa/
 # uk.ac.ebi.atlas.species.SpeciesPropertiesDao
 speciesPropertiesLocation = ${data.files.location}/species/
 
-# uk.ac.ebi.atlas.solr.admin.BioentityIndexAdmin
+# uk.ac.ebi.atlas.solr.bioentities.admin.BioentityIndexAdmin
 bioentity.properties = ${data.files.location}/bioentity_properties/
 
 # Ontology traders
